rust: put the logger into a dedicated module
