+++ rust/examples/collector.rs
@@ -16,8 +16,9 @@ use std::str::FromStr;
 use chrono::Local;
 use clap::{Arg, Command};
 use ini::Ini;
+use libits::client::configuration::Configuration;
 use libits::client::configuration::configuration_error::ConfigurationError;
-use libits::client::configuration::{Configuration, create_stdout_logger};
+use libits::client::logger::create_stdout_logger;
 use libits::transport::mqtt::mqtt_client::MqttClient;
 use libits::transport::mqtt::mqtt_router::MqttRouter;
 use libits::transport::mqtt::str_topic::StrTopic;
